+++ src/database/defines.h
@@ -545,7 +545,7 @@ public:
 		return this->domain_income_unit;
 	}
 
-	int get_domain_income_unit_value() const;
+	int64_t get_domain_income_unit_value() const;
 
 	const commodity_map<int> &get_province_level_commodity_costs_per_level() const
 	{
@@ -555,7 +555,7 @@ public:
 	int get_province_population_for_level(const int level) const;
 	int get_province_level_for_population(const int population) const;
 	const dice &get_province_taxation_for_level(const int level) const;
-	int get_domain_maintenance_cost_for_domain_size(const int domain_size) const;
+	int64_t get_domain_maintenance_cost_for_domain_size(const int domain_size) const;
 
 	const std::vector<std::unique_ptr<battle_resolution_table>> &get_battle_resolution_tables() const
 	{
@@ -664,7 +664,7 @@ private:
 	commodity_map<int> province_level_commodity_costs_per_level;
 	std::map<int, int> province_population_per_level;
 	std::map<int, dice> province_taxation_per_level;
-	std::map<int, int> domain_maintenance_cost_per_domain_size;
+	std::map<int, int64_t> domain_maintenance_cost_per_domain_size;
 	std::vector<std::unique_ptr<battle_resolution_table>> battle_resolution_tables;
 	std::map<terrain_adjacency, std::vector<int>> river_adjacency_subtiles;
 	std::map<terrain_adjacency, int> rivermouth_adjacency_tiles;
