+++ src/economy/expense_transaction.h
@@ -11,7 +11,7 @@ class expense_transaction final : public transaction
 	Q_OBJECT
 
 public:
-	explicit expense_transaction(const expense_transaction_type type, const int amount, const object_variant &object, const int object_quantity, const metternich::domain *domain)
+	explicit expense_transaction(const expense_transaction_type type, const int64_t amount, const object_variant &object, const int64_t object_quantity, const metternich::domain *domain)
 		: transaction(amount, object, object_quantity, domain), type(type)
 	{
 	}
