+++ src/infrastructure/wonder.h
@@ -80,12 +80,12 @@ public:
 
 	int get_wealth_cost_for_country(const domain *domain) const;
 
-	const commodity_map<int> &get_commodity_costs() const
+	const commodity_map<int64_t> &get_commodity_costs() const
 	{
 		return this->commodity_costs;
 	}
 
-	commodity_map<int> get_commodity_costs_for_country(const domain *domain) const;
+	commodity_map<int64_t> get_commodity_costs_for_country(const domain *domain) const;
 
 	const factor<domain> *get_cost_factor() const
 	{
@@ -123,7 +123,7 @@ private:
 	technology *required_technology = nullptr;
 	technology *obsolescence_technology = nullptr;
 	int wealth_cost = 0;
-	commodity_map<int> commodity_costs;
+	commodity_map<int64_t> commodity_costs;
 	std::unique_ptr<const factor<domain>> cost_factor;
 	std::unique_ptr<const and_condition<domain>> conditions;
 	std::unique_ptr<const and_condition<province>> province_conditions;
